fix: address CodeRabbit review on PR #28
- pack_repository: `topicsCount.clamp(1, 6)` returns `num`, cast to
  `int` via `.toInt()` so `List<Topic>.generate` typechecks.
- package_card: lift the reaction row (like/dislike/bookmark) out of
  `router.Link` — nesting `<button>` inside `<a>` is invalid HTML and
  a reaction click was bubbling into pack navigation.
- pack_score_panel: drop `aria-hidden` from the persistent scoreboard
  container. On desktop the panel is always visible via CSS, so
  hiding it from assistive tech misrepresented the state.
- pack_score_panel: keep the selected player pinned to the same
  entry when an earlier player is removed (decrement index).
- pack_details_page: wrap the clipboard write in try/catch and show
  a distinct failure toast if the sync call throws. Async promise
  rejection is still opaque without `dart:js_interop`; leave a TODO.

+++ sidb/lib/presentation/features/pack/repository/pack_repository.dart
@@ -83,7 +83,7 @@ class ApiPackRepository implements PackRepository {
       'Пакет собран из авторских вопросов, отыгранных на профильных турнирах. Ниже — темы и вопросы; ответ раскрывается по клику.';
 
   static List<Topic> _mockTopics(Pack base) {
-    final topicCount = base.topicsCount.clamp(1, 6);
+    final topicCount = base.topicsCount.clamp(1, 6).toInt();
     return List<Topic>.generate(topicCount, (i) {
       final index = i + 1;
       return Topic(
